Improve: users APIをuser.ts内に移植
